+++ workers/worker-jvm/src/main/java/com/openregex/engines/JvmStandard.java
@@ -60,6 +60,7 @@ public class JvmStandard {
                                     new CheatSheetItem("*", "0 or more times (greedy)"),
                                     new CheatSheetItem("+", "1 or more times (greedy)"),
                                     new CheatSheetItem("?", "0 or 1 time (greedy)"),
+                                    new CheatSheetItem("{m}", "Exactly m times"),
                                     new CheatSheetItem("{m,n}", "Between m and n times (greedy)"),
                                     new CheatSheetItem("*?", "0 or more times (lazy)"),
                                     new CheatSheetItem("+?", "1 or more times (lazy)"),